refactor(AboutDialog): consolidate dialog files
