+++ client/src/dialogs/About/AboutDialog.scss
