+++ client/src/dialogs/About/AboutDialog.tsx
@@ -1,15 +1,15 @@
 import React, { memo } from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { useSelector } from '../store/hooks'
-import ExternalLink from '../ui/ExternalLink'
-import logo from '../../images/logo_horizontal.svg'
-import numoLogo from '../../images/sponsors/numo.svg'
-import cfalogo from '../../images/sponsors/codeforamerica.png'
-import mozlogo from '../../images/sponsors/mozilla.svg'
-import Credits from './About/Credits'
-import SocialLinks from './About/SocialLinks'
-import Dialog from './Dialog'
+import { useSelector } from '~/src/store/hooks'
+import ExternalLink from '~/src/ui/ExternalLink'
+import logo from '~/images/logo_horizontal.svg'
+import numoLogo from '~/images/sponsors/numo.svg'
+import cfalogo from '~/images/sponsors/codeforamerica.png'
+import mozlogo from '~/images/sponsors/mozilla.svg'
+import Dialog from '../Dialog'
+import Credits from './Credits'
+import SocialLinks from './SocialLinks'
 import './AboutDialog.scss'
 
 function AboutDialog (): React.ReactElement {
